+++ source/miniiptv/telemetry_log.c
@@ -118,7 +118,8 @@ int miniiptv_telemetry_log_open(const char *path, const char *version,
         "rebuffer_wait_limit_ms,lag_segments,"
         "applied_lag_segments,profile_hit,segment_ms,download_ms,gap_ms,"
         "jitter_ms,rebuffering,recent_underruns,"
-        "total_underruns,global_underruns,downloaded_segments,no_new_streak,"
+        "total_underruns,global_underruns,downloaded_segments,sequence_resyncs,"
+        "no_new_streak,"
         "ring_min_bytes,ring_max_bytes,last_refill_ms,last_refill_commits,"
         "app_region_total_bytes,heap_total_bytes,heap_used_bytes,"
         "linear_total_bytes,linear_free_bytes,"
@@ -225,7 +226,7 @@ void miniiptv_telemetry_log_record(
         ",%" PRIu32 ",%" PRIu32
         ",%" PRIu32 ",%" PRIu32
         ",%" PRIu32 ",%" PRIu32 ",%u,%" PRIu32 ",%" PRIu32 ",%" PRIu64
-        ",%" PRIu64 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32
+        ",%" PRIu64 ",%" PRIu64 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32
         ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32
         ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32
         ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32 ",%" PRIu32
@@ -242,7 +243,7 @@ void miniiptv_telemetry_log_record(
         sample->commit_gap_deviation_ms, sample->rebuffering ? 1u : 0u,
         sample->recent_underruns, sample->total_underruns,
         sample->global_underruns, sample->downloaded_segments,
-        sample->no_new_poll_streak, sample->ring_min_bytes,
+        sample->sequence_resyncs, sample->no_new_poll_streak, sample->ring_min_bytes,
         sample->ring_max_bytes, sample->last_refill_ms,
         sample->last_refill_commits, sample->app_region_total_bytes,
         sample->heap_total_bytes, sample->heap_used_bytes,